fix: move notesController to separate file
